Add a Command class
The class holds two kinds of parameters. One is for passing prameters
to subprocess.run().  The other is for printing, which can be used
to run on the command line via copy & paste. The main difference
is using quotation for a printing string which shall be handled as
one paramter in printing mode.